requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-seed==0.3.1
djangorestframework==3.14.0
Faker==19.6.0
idna==3.4
psycopg2==2.9.7
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
